pyproject.toml
[tool.poetry]
name = "srnn"
version = "1.0.0"
description = "srnn"
authors = ["developer <developer@mail.com>"]


[tool.black]
line-length = 79


[tool.isort]
multi_line_output=3
include_trailing_comma=true
use_parentheses=true
line_length=79


[tool.taskipy.tasks]
start = "python start.py"



[tool.poetry.dependencies]
python = "^3.7, <3.9"
python-dotenv = "^0.13.0"
requests = "^2.23.0"
numpy = "^1.18.4"



[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
black = "^19.10b0"
taskipy = "^1.2.1"
